Fix problems with destroy and forcible destroy functionality: - hold/release device in start/done routines, this will probably slow   down things a bit, but previous code was racy; - only release device if g_gate_destroy() failed - if it succeeded device   is dead and there is nothing to release; - various other changes which makes forcible destruction reliable.
MFC after:	3 days

+++ sys/geom/gate/g_gate.h
@@ -68,9 +68,9 @@
  */
 struct g_gate_softc {
 	int			 sc_unit;		/* P: (read-only) */
-	int16_t			 sc_ref;		/* P: g_gate_list_mtx */
+	int			 sc_ref;		/* P: g_gate_list_mtx */
 	struct g_provider	*sc_provider;		/* P: (read-only) */
-	uint32_t		 sc_flags;		/* P: (read-only) */
+	uint32_t		 sc_flags;		/* P: sc_queue_mtx */
 
 	struct bio_queue_head	 sc_inqueue;		/* P: sc_queue_mtx */
 	struct bio_queue_head	 sc_outqueue;		/* P: sc_queue_mtx */
